Items, Crafting, Blueprints
Now you can pick up items with F and open your inventory with I to view the crafting bench and then craft items. Models and sprite icons need to be done still. You need to pick up the blueprints to be able to craft an Axe and Medkit, other one is for testing. Then you caan equip. Try it out!